authentication core

+++ kotlin/src/main/kotlin/com/job/library/common/money/Money.kt
@@ -1,8 +1,6 @@
 package com.job.library.common.money
 
-import java.math.BigInteger
-
 data class Money(
-    val amount: BigInteger,
+    val amount: Long,
     val currency: Currency,
 )
